Fixed everything, hopefully
